+++ src/Koffee.Tests/MainSearchTests_Find.fs
@@ -9,74 +9,73 @@ let items names =
     let toItem name = Item.Basic Path.Root name Folder
     names |> Seq.map toItem |> Seq.toList
 
-let testModel cursorStart =
-    { testModel with
-        Items = "alice,bob,charlie,crystal,apple,cherry".Split(',') |> items
-        Cursor = cursorStart
-    }
-
-let assertEqualExceptCursor expected actual =
-    assertAreEqualWith expected actual (ignoreMembers ["Cursor"; "Items"; "InputText"])
-
-let doFind next repeat prefix cursorStart =
+let doFind next repeat findText cursorStart =
     let model =
-        { testModel cursorStart with
-            LastFind = Some prefix
+        { testModel with
+            Items = "alice,bob,charlie,crystal,apple,cherry".Split(',') |> items
+            Cursor = cursorStart
             RepeatCommand = repeat
+            InputError = Some "old error"
+            InputText = if not next then findText else ""
+            LastFind = if next then Some findText else None
         }
 
     let actual =
         if next then
             Search.findNext model
         else
-            Search.find prefix model
+            Search.find model
 
     let expected =
         { model with
-            LastFind = Some prefix
-            Status = if next then Some (MainStatus.find prefix (repeat |? 1)) else None
+            LastFind = findText |> Option.ofString
+            Status = if next then Some (MainStatus.find findText (repeat |? 1)) else None
         }
-    assertEqualExceptCursor expected actual
-    actual.Cursor
+    assertAreEqualWith expected actual (ignoreMembers ["Cursor"; "InputError"])
+    (actual.Cursor, actual.InputError)
 
 let find = doFind false None
 let findNext = doFind true None
 let findNextNth count = doFind true (Some count)
 
 [<Test>]
-let ``Find that matches nothing should not change the cursor``() =
-    find "d" 1 |> shouldEqual 1
+let ``Find empty string clears error``() =
+    find "" 0 |> shouldEqual (0, None)
+
+[<Test>]
+let ``Find that matches nothing should show error and not change the cursor``() =
+    find "d" 1 |> shouldEqual (1, Some (ErrorMessages.findFailed "d"))
 
 [<Test>]
 let ``Find that matches only the current item should not change the cursor``() =
-    find "b" 1 |> shouldEqual 1
+    find "b" 1 |> shouldEqual (1, None)
 
 [<Test>]
 let ``Find that matches the current and next item should not change the cursor``() =
-    find "c" 2 |> shouldEqual 2
+    find "c" 2 |> shouldEqual (2, None)
 
 [<Test>]
-let ``Find that matches a item wrapping around should set the cursor to the that index``() =
-    find "b" 2 |> shouldEqual 1
+let ``Find that matches an item wrapping around should set the cursor to the that index``() =
+    find "b" 2 |> shouldEqual (1, None)
 
 [<Test>]
-let ``Find next that matches nothing should not change the cursor``() =
-    findNext "d" 1 |> shouldEqual 1
+let ``Find next that matches nothing should show error and not change the cursor``() =
+    findNext "d" 1 |> shouldEqual (1, Some (ErrorMessages.findFailed "d"))
 
 [<Test>]
 let ``Find next that matches only the current item should not change the cursor``() =
-    findNext "b" 1 |> shouldEqual 1
+    findNext "b" 1 |> shouldEqual (1, None)
 
 [<Test>]
 let ``Find next that matches the current and next item should set the cursor to the next index``() =
-    findNext "c" 2 |> shouldEqual 3
+    findNext "c" 2 |> shouldEqual (3, None)
 
 [<Test>]
-let ``Find next that matches a item wrapping around should set the cursor to the that index``() =
-    findNext "b" 2 |> shouldEqual 1
+let ``Find next that matches an item wrapping around should set the cursor to the that index``() =
+    findNext "b" 2 |> shouldEqual (1, None)
 
 [<TestCase(1, 2)>]
 [<TestCase(2, 3)>]
 [<TestCase(3, 5)>]
 let ``Find next nth that matches`` nth expected =
-    findNextNth nth "c" 1 |> shouldEqual expected
+    findNextNth nth "c" 1 |> shouldEqual (expected, None)